feat: implement QuizReviewRestController for managing quiz reviews
